Include assumptions in ->dimacs
ClauseStorage->dimacs forgot to add the assumptions to the CNF file.
The quick and easy way to fix this was to make a copy of the clauses
and append the assumptions as extra clauses to this temporary copy.
There is some opportunity for reducing memory consumption here, when
formulas are large.