Provide incremental SAT solver
This is an extra package outside the main ManySAT.pm because it uses a
different storage format (the CaDiCaL solver object which is refinable).

In the context of CInet, a general one-off satisfiability problem should
use ManySAT.pm. This is the canonical solver. But a filter for some quasi-
axiomatic property like orientability, where the same formula is checked
against varying assumptions 40 million times, should use Incremental.pm.