Use ref instead of reftype
I actually want to detect *unblessed* array references. This changes
comes with the documented drawbacks/ambiguities of ref but it fixes
the problem that ManySAT::Base->read does not mistake a Path::Tiny
(which is apparently a blessed arrayref) for an array of clauses.

This also means that clauses can never be arrayrefs that happen to
be blessed into something, which is probably okay.

Since ref, unlike reftype, returns empty strings instead of undefs,
I can remove all `no warnings "uninitialized"`s.